Correctly convert a decimal value to int base

+++ SBRW.Launcher.Core.Downloader/LZMA_/EventArg_/Download_Data_Progress_EventArgs.cs
@@ -26,7 +26,7 @@ namespace SBRW.Launcher.Core.Downloader.LZMA_.EventArg_
         /// <summary>
         /// 
         /// </summary>
-        public long Bytes_Received_Over_Total { get; internal set; }
+        public decimal Bytes_Received_Over_Total { get; internal set; }
         /// <summary>
         /// 
         /// </summary>
@@ -40,7 +40,7 @@ namespace SBRW.Launcher.Core.Downloader.LZMA_.EventArg_
         /// <param name="Calulated_Percentage"></param>
         /// <param name="Received_File_Name"></param>
         /// <param name="Received_Start_Time"></param>
-        public Download_Data_Progress_EventArgs(long Received_Bytes_Current, long Received_Bytes_Total, long Calulated_Current_Divide_Total, int Calulated_Percentage, string Received_File_Name, DateTime Received_Start_Time)
+        public Download_Data_Progress_EventArgs(long Received_Bytes_Current, long Received_Bytes_Total, decimal Calulated_Current_Divide_Total, int Calulated_Percentage, string Received_File_Name, DateTime Received_Start_Time)
         {
             this.Bytes_To_Receive_Total = Received_Bytes_Total;
             this.Bytes_Received = Received_Bytes_Current;
